pyproject.toml
[project]
name = "vae"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "dlutils>=0.0.12",
    "gdown>=5.2.0",
    "imageio>=2.37.0",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "scikit-learn>=1.6.1",
    "scipy>=1.13.1",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "yacs>=0.1.8",
]
